change uint8_t to unsigned for bit fields

+++ include/gbcpu.h
@@ -5,11 +5,11 @@
 
 typedef struct
 {
-    uint8_t _unused : 4; // bits 0-3, always 0
-    uint8_t c : 1;       // bit 4
-    uint8_t h : 1;       // bit 5
-    uint8_t n : 1;       // bit 6
-    uint8_t z : 1;       // bit 7
+    unsigned _unused : 4; // bits 0-3, always 0
+    unsigned c : 1;       // bit 4
+    unsigned h : 1;       // bit 5
+    unsigned n : 1;       // bit 6
+    unsigned z : 1;       // bit 7
 } __attribute__((packed)) gbcpu_flags_t;
 
 typedef struct
